+++ tests/CallMethodTest.php
@@ -15,6 +15,14 @@ final class CallMethodTest extends PHPUnit_Framework_TestCase
         $this->assertSame(123, $fn(new CallMethodExample(123)));
     }
 
+    /** @test */
+    public function it_calls_the_method_with_arguments()
+    {
+        $fn = T\callMethod('returnArgument', 'argument value');
+
+        $this->assertEquals('argument value', $fn(new CallMethodExample()));
+    }
+
     /**
      * @test
      * @expectedException \InvalidArgumentException
@@ -29,7 +37,7 @@ class CallMethodExample
 {
     private $value;
 
-    public function __construct($value)
+    public function __construct($value = null)
     {
         $this->value = $value;
     }
@@ -37,5 +45,10 @@ class CallMethodExample
     public function getValue()
     {
         return $this->value;
+    }
+
+    public function returnArgument($argument)
+    {
+        return $argument;
     }
 }